+++ src/components/profile/vehicleForm/VehicleForm.jsx
@@ -1,5 +1,5 @@
 import { useForm } from 'react-hook-form';
-import { Button } from '../ui/button';
+import { Button } from '../../ui/button';
 import {
   Form,
   FormControl,
@@ -7,10 +7,10 @@ import {
   FormItem,
   FormLabel,
   FormMessage
-} from '../ui/form';
-import { Input } from '../ui/input';
+} from '../../ui/form';
+import { Input } from '../../ui/input';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { SheetClose } from '../ui/sheet';
+import { SheetClose } from '../../ui/sheet';
 import { cn } from '@/lib/utils';
 
 const VehicleForm = ({
